+++ cmd/discovery/connector_kubernetes.go
@@ -7,7 +7,7 @@ import (
 	"github.com/jaku01/caddyservicediscovery/internal/kubernetes"
 )
 
-// newConnector returns a Kubernetes connector when built with the 'kubernetes' tag.
+// Returns a Kubernetes connector when built with the 'kubernetes' tag.
 func newProviderConnector() (caddy.ProviderConnector, error) {
 	return kubernetes.NewKubernetesConnector()
 }
